requirements.txt
distlib==0.4.0
filelock==3.19.1
mysql-connector-python==9.4.0
platformdirs==4.4.0
virtualenv==20.34.0
